requirements.txt
mfrc522==0.0.7
RPi.GPIO==0.7.1
RPLCD==1.3.1
smbus==1.1.post2
spidev==3.6